requirements.txt
charset-normalizer==3.3.2
idna==3.4
mysql-connector==2.2.9
pip==23.3.1
requests==2.31.0
urllib3==2.1.0